The goal in this assignment was filtering in the spatial domain as well as in the frequency domain. Laplacian Blending using Image Pyramids is a very good intro to working and thinking in frequencies, and Deconvolution is a neat trick.
Tasks:
Perform Histogram Equalization on the given input image.
Perform Low-Pass, High-Pass and Deconvolution on the given input image.
Perform Laplacian Blending on the two input images (blend them together).

Histogram Equalization
Refer to the readings on @43, particularly to Szeliski's section 3.4.1, and within it to eqn 3.9.
Getting the histogram of a grayscale image is incredibly easy (Python):

Low-Pass, High-Pass and Deconvolution in the Frequency Domain
http://docs.opencv.org/master/de/dbc/tutorial_py_fourier_transform.html
http://docs.opencv.org/2.4/doc/tutorials/core/discrete_fourier_transform/discrete_fourier_transform.html

LPF: mask a 20x20 window of the center of the FT image (the low frequencies).
HPF: just reverse the mask.

Deconvolution: apply a gaussian kernel (gk) to your input image (in the FD/FFT):

Laplacian Pyramid Blending
http://docs.opencv.org/3.2.0/dc/dff/tutorial_py_pyramids.html